Fixed template instantiation problems.
My template instantiation system needed serious modification for when I deprecated a class and provided a new implementation. The system failed to instantiate the new implementations. I discovered this when attempting to backport the IntpQuadraticNonuniform2 class (the Cendes-Wong algorithm) from GTL to GTE. The GTL code has been thoroughly unit tested and verified with Mathematica. There are no sample applications that exercise those new implementations, which was clear when the code would not compile---a serious oversight on my part. I have made changes to various classes to get the template instantiation to compile the new code. I will post the backported GTL code for IntpQuadraticNonuniform2 after I unit test the derived classes IntpSphere2 and IntpVectorField2.

+++ GTE/Mathematics/IntpLinearNonuniform2.h
@@ -3,7 +3,7 @@
 // Distributed under the Boost Software License, Version 1.0.
 // https://www.boost.org/LICENSE_1_0.txt
 // https://www.geometrictools.com/License/Boost/LICENSE_1_0.txt
-// Version: 6.0.2023.08.08
+// Version: 6.0.2024.03.12
 
 #pragma once
 
@@ -43,7 +43,7 @@ namespace gte
         // which case the interpolation is valid.
         bool operator()(Vector2<Real> const& P, Real& F) const
         {
-            int32_t t = mMesh->GetContainingTriangle(P);
+            int32_t t = static_cast<int32_t>(mMesh->GetContainingTriangle(P));
             if (t == -1)
             {
                 // The point is outside the triangulation.
